Add files via upload
A calculator which preforms addition, subtraction, division or multiplication on two user selected numbers ranging from 0-9.
in phase 1, pressing the A button will advance the number by one, and conversely pressing the B button will decrease the number by one. Pressing A + B will then take the user to an operation select, where pressing A or B will cycle through the available options. Yet another press of A+B will progress the program to the second numbers selection, and a last push of A+B will display the result.